move integration tests to correct folder
