+++ test/integration/graphql/business.test.js
@@ -2,17 +2,28 @@ import { deepEqual } from 'assert'
 
 import { graphql } from 'graphql'
 
-import { schema } from '../../app/graphql/server.js'
-import { context } from '../../app/graphql/context.js'
+import { schema } from '../../../app/graphql/server.js'
+import { context } from '../../../app/graphql/context.js'
 
-import mockServer from '../../mocks/server.js'
+import mockServer from '../../../mocks/server.js'
 
-import { transformOrganisationToBusiness } from '../../app/transformers/rural-payments-portal/business.js'
-import { organisation as organisationFixture } from '../../mocks/fixtures/organisation.js'
+import { transformOrganisationToBusiness } from '../../../app/transformers/rural-payments-portal/business.js'
+import { organisation as organisationFixture } from '../../../mocks/fixtures/organisation.js'
 
-describe('Query.business', () => {
-  before(mockServer.start)
-  after(mockServer.stop)
+const ruralPaymentsPortalApiUrl = process.env.RURAL_PAYMENTS_PORTAL_API_URL
+const ruralPaymentsPortalEmail = process.env.RURAL_PAYMENTS_PORTAL_EMAIL
+
+describe('Query.customer', () => {
+  before(async () => {
+    const mockUrl = await mockServer.start()
+    process.env.RURAL_PAYMENTS_PORTAL_API_URL = `${mockUrl}/rpp/`
+    process.env.RURAL_PAYMENTS_PORTAL_EMAIL = 'test-email'
+  })
+  after(async () => {
+    process.env.RURAL_PAYMENTS_PORTAL_API_URL = ruralPaymentsPortalApiUrl
+    process.env.RURAL_PAYMENTS_PORTAL_EMAIL = ruralPaymentsPortalEmail
+    await mockServer.stop()
+  })
 
   it('should return business data', async () => {
     await mockServer.selectBase('base')
