+++ test/integration/graphql/land.test.js
@@ -2,8 +2,8 @@ import { deepEqual } from 'assert'
 
 import { graphql } from 'graphql'
 
-import { schema } from '../../app/graphql/server.js'
-import { context } from '../../app/graphql/context.js'
+import { schema } from '../../../app/graphql/server.js'
+import { context } from '../../../app/graphql/context.js'
 
 describe('Query.land', () => {
   it('should return land data', async () => {
@@ -70,5 +70,5 @@ describe('Query.land', () => {
         }
       }
     })
-  })
+  }).timeout(10000)
 })
